requirements.txt
Django
djangorestframework
django-dotenv
django-redis
psycopg2-binary
requests
djangorestframework-simplejwt

